Fix #189, replaced `sys.exit` with the stronger `os._exit`
The SDTmonitor worker processes now terminate properly. The `os._sys`
function actually forces the process to terminate immediately without
performing a proper cleanup procedure. I believe the issue causing the
process not no terminate properly with `sys.exit` is somehow related to
the cleanup procedure and/or the garbage collection. Since this happens
randomly, I haven't been able to find a proper solution other than this
workaround (which, in the end, seems to not introduce any other issue).